refactor: (피드백반영) Number Range Validation 로직 위치 변경 및 테스트 추가 - Lotto > Number
