+++ src/main/java/lotto/domain/Lotto.java
@@ -9,12 +9,9 @@ public class Lotto {
     public static final int LOTTO_PRICE = 1_000;
 
     public static final int LOTTO_NUMBERS_SIZE = 6;
-    public static final int SAVED_MIN_NUMBER = 1;
-    public static final int SAVED_MAX_NUMBER = 45;
 
     private static final String LOTTO_NUMBERS_INVALID_SIZE_ERROR_MESSAGE = "로또 번호의 개수는 6개가 들어와야 한다.";
     private static final String LOTTO_NUMBERS_INVALID_DUPLICATE_ERROR_MESSAGE = "로또의 번호는 중복되어 저장될 수 없다.";
-    private static final String LOTTO_NUMBERS_INVALID_RANGE_ERROR_MESSAGE = "로또 번호는 1이상 45이하의 수만 들어올 수 있다.";
     private static final String LOTTO_NUMBERS_INVALID_ASC_ERROR_MESSAGE = "로또 번호는 오름차순으로 입력되어있어야 한다.";
 
     private final Numbers numbers;
@@ -22,7 +19,6 @@ public class Lotto {
     public Lotto(List<Number> numbers) {
         checkLottoNumbersSize(numbers);
         checkDuplicatedNumber(numbers);
-        checkNumberRange(numbers);
         checkNumberAcs(numbers);
 
         this.numbers = new Numbers(numbers);
@@ -40,22 +36,12 @@ public class Lotto {
         }
     }
 
-    private static void checkNumberRange(List<Number> numbers) {
-        if (numbers.stream().map(Number::value).filter(Lotto::isNumberRange).count() != LOTTO_NUMBERS_SIZE) {
-            throw new IllegalArgumentException(LOTTO_NUMBERS_INVALID_RANGE_ERROR_MESSAGE);
-        }
-    }
-
     private static void checkNumberAcs(List<Number> numbers) {
         List<Number> compare = new ArrayList<>(numbers);
         Collections.sort(compare);
         if (!numbers.equals(compare)) {
             throw new IllegalArgumentException(LOTTO_NUMBERS_INVALID_ASC_ERROR_MESSAGE);
         }
-    }
-
-    private static boolean isNumberRange(int num) {
-        return num >= SAVED_MIN_NUMBER && num <= SAVED_MAX_NUMBER;
     }
 
     public Numbers numbers() {
