+++ src/test/java/lotto/domain/LottoTest.java
@@ -44,22 +44,6 @@ class LottoTest {
     }
 
     @Test
-    @DisplayName("로또 번호에 1이상 45이하의 수가 들어오지 않으면 Exception이 발생되어야 한다.")
-    void lottoSaveNumberRangeTest() {
-
-        // given
-        List<Number> input = Arrays.asList(1, 2, 3, 4, 5, 76)
-            .stream()
-            .map(Number::new)
-            .collect(toList());
-
-        // when & then
-        assertThatExceptionOfType(IllegalArgumentException.class)
-            .isThrownBy(() -> new Lotto(input))
-            .withMessageMatching("로또 번호는 1이상 45이하의 수만 들어올 수 있다.");
-    }
-
-    @Test
     @DisplayName("로또의 번호가 오름차순으로 입력되지 않으면 Exception이 발생되어야 한다.")
     void lottoSaveNumberAscTest() {
